Initiated SubscriberController and subscriber.js. Moved some code to the new files. Also did some refactorizing
