+++ src/main/resources/templates/fragments/indexLoggedIn.html
@@ -11,7 +11,7 @@
                 <li role="presentation" class="active"><a href="#main" data-toggle="tab">Main</a></li>
                 <li role="presentation"><a href="#topics" data-toggle="tab">Topics</a></li>
                 <li role="presentation"><a href="#subscribers" data-toggle="tab">Subscribers</a></li>
-                <li role="presentation"><a href="#stats" data-toggle="tab">Statistics</a></li>
+                <li role="presentation"><a href="#statistics" data-toggle="tab">Statistics</a></li>
                 <li role="presentation"><a href="#log" data-toggle="tab">Logs</a></li>
                 <li role="presentation"><a href="#config" data-toggle="tab">Configuration</a></li>
             </ul>
