+++ src/main/resources/templates/fragments/panes/statspane.html
@@ -6,7 +6,7 @@
 </head>
 <body>
 <th:block layout:fragment="statspane">
-    <div class="tab-pane" id="stats">
+    <div class="tab-pane" id="statistics">
         <div class="row">
             <div class="col-md-8">
                 <!-- All protocol stats will be inserted here -->
