+++ src/main/resources/templates/fragments/panes/topicspane.html
@@ -28,9 +28,9 @@
                             <thead>
                                 <tr>
                                     <th>Topic</th>
+                                    <th>Subscribers</th>
                                     <th>Root</th>
                                     <th>Leaf</th>
-                                    <th>Subscribers</th>
                                     <th>Actions</th>
                                 </tr>
                             </thead>
